change pointer type: int -> size_t

+++ libds.c
@@ -27,7 +27,7 @@ void stack_destroy(stack_t *stack)
 
 void stack_push(stack_t *stack, void *item)
 {
-	if (stack->pointer >= (int)stack->capacity)
+	if (stack->pointer >= stack->capacity)
 	{
 		size_t capacity = stack->capacity + 10;
 		void **values_new = realloc(stack->values, capacity * sizeof(void *));
